break up expansion.py to a module
