+++ src/gtc/dace/__init__.py
@@ -17,7 +17,7 @@
 # need to import expansion so that implementation of expansion is registered
 from typing import Any, List
 
-from .expansion import (  # noqa: F401 '.expansion.StencilComputationExpansion' imported but unused
+from .expansion.expansion import (  # noqa: F401 '.expansion.StencilComputationExpansion' imported but unused
     StencilComputationExpansion,
 )
